refactor($location): change package structure
패키지 구성을 변경하여 전체적인 이동이 일어났습니다.
