+++ studycafe/src/main/java/com/scascanner/studycafe/domain/entity/ReservationStatus.java
@@ -1,4 +1,4 @@
-package com.scascanner.studycafe.domain;
+package com.scascanner.studycafe.domain.entity;
 
 public enum ReservationStatus {
     RESERVED,CANCELED
