+++ studycafe/src/main/java/com/scascanner/studycafe/web/login/controller/UserApiController.java
@@ -1,10 +1,8 @@
-package com.scascanner.studycafe.login.controller;
+package com.scascanner.studycafe.web.login.controller;
 
-import com.scascanner.studycafe.login.dto.UserForm;
-import com.scascanner.studycafe.login.dto.UserInfoDto;
-import com.scascanner.studycafe.login.service.UserService;
-import lombok.AllArgsConstructor;
-import lombok.Data;
+import com.scascanner.studycafe.web.login.dto.UserForm;
+import com.scascanner.studycafe.web.login.dto.UserInfoDto;
+import com.scascanner.studycafe.web.login.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
